+++ src/views/scores/ScoresTapestry.vue
@@ -1,28 +1,62 @@
 <template>
   <v-container>
-    <h1 class="text-h2 text-center">Tapestry scores</h1>
+    <h1 class="text-h2 text-center mb-5">Tapestry scores</h1>
 
-    <TapestryScoresTable
-      :civilizations="tapestryCivilizations"
-      :plays="tapestryPlays"
-    />
+    <v-tabs v-model="tab" align-tabs="center">
+      <v-tab :value="1">History</v-tab>
+      <v-tab :value="2">Civilizations</v-tab>
+      <v-tab :value="3">Scenarios</v-tab>
+    </v-tabs>
+
+    <v-tabs-items v-model="tab">
+      <v-tab-item>
+        <TapestryScoresByDate
+          :civilizations="tapestryCivilizations"
+          :plays="tapestryPlays"
+        />
+      </v-tab-item>
+      <v-tab-item>
+        <TapestryScoresByCivilization
+          :civilizations="tapestryCivilizations"
+          :plays="tapestryPlays"
+        />
+      </v-tab-item>
+      <v-tab-item>
+        <TapestryScoresByScenario
+          :civilizations="tapestryCivilizations"
+          :plays="tapestryPlays"
+          :campaigns="tapestryCampaigns"
+          :scenarios="tapestryScenarios"
+        />
+      </v-tab-item>
+    </v-tabs-items>
   </v-container>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
-import TapestryScoresTable from "@/components/TapestryScoresTable.vue";
+import TapestryScoresByCivilization from "@/components/TapestryScoresByCivilization.vue";
+import TapestryScoresByDate from "@/components/TapestryScoresByDate.vue";
+import TapestryScoresByScenario from "@/components/TapestryScoresByScenario.vue";
 import { mapState } from "vuex";
 
 export default Vue.extend({
   name: "ScoresTapestry",
-
+  data: () => ({
+    tab: null,
+  }),
   components: {
-    TapestryScoresTable,
+    TapestryScoresByCivilization,
+    TapestryScoresByDate,
+    TapestryScoresByScenario,
   },
-
   computed: {
-    ...mapState(["tapestryPlays", "tapestryCivilizations"]),
+    ...mapState([
+      "tapestryPlays",
+      "tapestryCivilizations",
+      "tapestryCampaigns",
+      "tapestryScenarios",
+    ]),
   },
 });
 </script>